requirements.txt
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-SocketIO==5.3.5
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
requests==2.31.0
texttable==1.6.7
urllib3==2.0.4
Werkzeug==2.3.7
